refactor structure project. add linter. add unit-test.
