+++ .editorconfig
@@ -1,0 +1,5 @@
+[*.js]
+indent_style = space
+indent_size = 2
+trim_trailing_whitespace = true
+insert_final_newline = true
