+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+   "xo.enable": true,
+   "eslint.enable": false
+}
